Don't use forward typedefs to avoid break on older GCC
Older versions of gcc (4.5.2 in particular) don't understand a
forward typedef and complain a type is redefined when it sees the
actual definition.

Change OQS_KEX to refer to itself as "struct OQS_KEX" in its definition
to remove the need for this forward definition and avoid the redefinition
error on older versions of gcc.