Fix for the Zephyr CI tests
The Zephyr Github CI Actions used a complete Zephyr installation to
build and run the test code. The resulting size of the data has been up
to several GBs, overfilling the Github runner storage space.

With this patch of the Zephyr workflow, only a minimal Zephyr
installation is created to run the tests, resulting in a much smaller
space footprint.

Furthermore, the timeout of the Signature test has been increased, as
sometimes the tests take longer than the prior 10 minutes.

Signed-off-by: Tobias Frauenschläger <t.frauenschlaeger@me.com>