add compile_commands.json to .gitignore
We set CMAKE_EXPORT_COMPILE_COMMANDS=ON so the compilation database is
automatically exported to the build directory. However, many language
servers expect this file to be in the project root directory, so a common
post-build step is to:

```shell
$ ln -s <build-directory>/compile_commands.json
```

This PR enables developers to do this without having to worry about
accidentally commiting a symlink file.

Signed-off-by: carson radtke <nosrac925@gmail.com>