Adjust constan-time test exception for mlkem-native [extended tests]
The constant-time tests have been failing for mlkem-native as check_sk changed it's name and is no no longer recognized as an exception.
This function processes exclusively public data in the secret key and it's, hence, okay to branch both inside the function and depending on the return value.

This commit renames the function in the constant_time exceptions file.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>